Ability to Scan QR containing a URL to a zip download
git-svn-id: http://cyanogen-updater.googlecode.com/svn/trunk@113 fa1c2e9a-81d2-11de-92f0-6bbadbc12360

+++ src/cmupdater/service/UpdateDownloaderService.java
@@ -72,14 +72,14 @@ import cmupdater.utils.Preferences;
 
 
 public class UpdateDownloaderService extends Service {
-	
+
 	//public static UpdateDownloaderService INSTANCE;
-	
+
 	private static final String TAG = "UpdateDownloader";
-	
+
 	public static final String KEY_REQUEST = "cmupdater.request";
 	public static final String KEY_UPDATE_INFO = "cmupdater.updateInfo";
-	
+
 	public static final int REQUEST_DOWNLOAD_UPDATE = 1;
 
 	private final BroadcastReceiver mConnectivityChangesReceiver = new BroadcastReceiver() {
@@ -109,8 +109,8 @@ public class UpdateDownloaderService extends Service {
 			}
 		}
 	};*/
-	
-	
+
+
 	private Looper mServiceLooper;
 	private ServiceHandler mServiceHandler;
 	private HandlerThread mHandlerThread;
@@ -125,28 +125,28 @@ public class UpdateDownloaderService extends Service {
 	private boolean mMirrorNameUpdated;
 	private String mMirrorName;
 	private String mFileName;
-	
+
 	private boolean mDownloading = false;
 	private UpdateInfo mCurrentUpdate;
 	private ConnectivityManager mConnectivityManager;
 	private IntentFilter mConectivityManagerIntentFilter;
-    private final IBinder mBinder = new LocalBinder();
-	
+	private final IBinder mBinder = new LocalBinder();
+
 	private static IUpdateProcessInfo UPDATE_PROCESS_INFO;
 
 	private WifiLock mWifiLock;
 	private WifiManager mWifiManager;
-	
+
 	private String mUpdateFolder;
 	private String mDownlaodedMD5;
-	
-	 public class LocalBinder extends Binder {
-        public UpdateDownloaderService getService() {
-            return UpdateDownloaderService.this;
-        }
-	 }
 
-	
+	public class LocalBinder extends Binder {
+		public UpdateDownloaderService getService() {
+			return UpdateDownloaderService.this;
+		}
+	}
+
+
 	private final class ServiceHandler extends Handler {
 
 		public ServiceHandler(Looper looper) {
@@ -155,57 +155,57 @@ public class UpdateDownloaderService extends Service {
 
 		@Override
 		public void handleMessage(Message msg) {
-            Bundle arguments = (Bundle)msg.obj;
-            
-            int request = arguments.getInt(KEY_REQUEST); 
-            switch(request) {
-            	case REQUEST_DOWNLOAD_UPDATE:
-            		mDownloading = true;
-            		try {
-						UpdateInfo ui = mCurrentUpdate = (UpdateInfo) arguments.getSerializable(KEY_UPDATE_INFO);
-	            		File downloadedUpdate = checkForConnectionAndUpdate(ui);
-	            		notifyUser(ui, downloadedUpdate);
-            		} finally {
-            			mDownloading = false;
-            		}
-					break;
-            	default:
-            		Log.e(TAG, "Unknown request ID:" + request);
-            }
+			Bundle arguments = (Bundle)msg.obj;
 
-            Log.i("ServiceStartArguments", "Done with #" + msg.arg1);
-            stopSelf(msg.arg1);
+			int request = arguments.getInt(KEY_REQUEST); 
+			switch(request) {
+			case REQUEST_DOWNLOAD_UPDATE:
+				mDownloading = true;
+				try {
+					UpdateInfo ui = mCurrentUpdate = (UpdateInfo) arguments.getSerializable(KEY_UPDATE_INFO);
+					File downloadedUpdate = checkForConnectionAndUpdate(ui);
+					notifyUser(ui, downloadedUpdate);
+				} finally {
+					mDownloading = false;
+				}
+				break;
+			default:
+				Log.e(TAG, "Unknown request ID:" + request);
+			}
+
+			Log.i("ServiceStartArguments", "Done with #" + msg.arg1);
+			stopSelf(msg.arg1);
 		}
 	}
-	
+
 	public static void setUpdateProcessInfo(IUpdateProcessInfo iupi) {
 		UPDATE_PROCESS_INFO = iupi;
 	}
-	
+
 
 	@Override
 	public void onCreate() {
-        mNM = (NotificationManager)getSystemService(NOTIFICATION_SERVICE);
-        //mTelephonyManager = (TelephonyManager) getSystemService(TELEPHONY_SERVICE);
-        mConnectivityManager = (ConnectivityManager) getSystemService(CONNECTIVITY_SERVICE);
-        mConectivityManagerIntentFilter = new IntentFilter(ConnectivityManager.CONNECTIVITY_ACTION);
-        
+		mNM = (NotificationManager)getSystemService(NOTIFICATION_SERVICE);
+		//mTelephonyManager = (TelephonyManager) getSystemService(TELEPHONY_SERVICE);
+		mConnectivityManager = (ConnectivityManager) getSystemService(CONNECTIVITY_SERVICE);
+		mConectivityManagerIntentFilter = new IntentFilter(ConnectivityManager.CONNECTIVITY_ACTION);
+
 		mHandlerThread = new HandlerThread(TAG);
 		mHandlerThread.start();
-        mServiceLooper = mHandlerThread.getLooper();
-        mServiceHandler = new ServiceHandler(mServiceLooper);
-        //Resources res = getResources();
-		
+		mServiceLooper = mHandlerThread.getLooper();
+		mServiceHandler = new ServiceHandler(mServiceLooper);
+		//Resources res = getResources();
+
 		//String destFileName = res.getString(R.string.conf_update_file_name);
 		//mDestinationFile = new File(Environment.getExternalStorageDirectory(), destFileName);
 		//mDestinationMD5File = new File(Environment.getExternalStorageDirectory(), destFileName + ".md5sum");
 		mHttpClient = new DefaultHttpClient();
 		mMD5HttpClient = new DefaultHttpClient();
 		mRandom = new Random();
-		
+
 		mWifiManager = (WifiManager) getSystemService(WIFI_SERVICE);
 		mWifiLock = mWifiManager.createWifiLock("JF Updater");
-		
+
 		mUpdateFolder = Preferences.getPreferences(this).getUpdateFolder();
 	}
 
@@ -217,17 +217,17 @@ public class UpdateDownloaderService extends Service {
 				return;
 			}
 		}
-        Log.i(TAG, "Starting #" + startId + ": " + intent.getExtras());
-        
-        Message msg = mServiceHandler.obtainMessage();
-        msg.arg1 = startId;
-        msg.obj = intent.getExtras();
-        mServiceHandler.sendMessage(msg);
-        Log.d("ServiceStartArguments", "Sending: " + msg);
+		Log.i(TAG, "Starting #" + startId + ": " + intent.getExtras());
+
+		Message msg = mServiceHandler.obtainMessage();
+		msg.arg1 = startId;
+		msg.obj = intent.getExtras();
+		mServiceHandler.sendMessage(msg);
+		Log.d("ServiceStartArguments", "Sending: " + msg);
 	}
-	
-	
-	
+
+
+
 	private boolean isDataConnected() {
 		if (mConnectivityManager.getActiveNetworkInfo() == null)
 		{
@@ -238,12 +238,12 @@ public class UpdateDownloaderService extends Service {
 			return mConnectivityManager.getActiveNetworkInfo().isConnected();
 		}
 	}
-	
+
 	/**
 	private boolean isWifiNetwork() {
 		return mConnectivityManager.getActiveNetworkInfo().getType() == ConnectivityManager.TYPE_WIFI;
 	}
-	*/
+	 */
 
 	@Override
 	public void onDestroy() {
@@ -256,7 +256,7 @@ public class UpdateDownloaderService extends Service {
 	public IBinder onBind(Intent intent) {
 		return mBinder;
 	}
-	
+
 	/**
 	 * @return the downloading
 	 */
@@ -275,7 +275,7 @@ public class UpdateDownloaderService extends Service {
 
 	private File checkForConnectionAndUpdate(UpdateInfo updateToDownload) {
 		File downloadedFile;
-		
+
 		//wait for a data connection
 		while(!isDataConnected()) {
 			Log.d(TAG, "No data connection, waiting for a data connection");
@@ -288,9 +288,9 @@ public class UpdateDownloaderService extends Service {
 				}
 			}
 		}
-		
+
 		mWifiLock.acquire();
-		
+
 		try {
 			Log.i(TAG, "Downloading update...");
 			downloadedFile = downloadFile(updateToDownload);
@@ -301,7 +301,7 @@ public class UpdateDownloaderService extends Service {
 		} finally {
 			mWifiLock.release();
 		}
-		
+
 		return downloadedFile;
 	}
 
@@ -309,7 +309,7 @@ public class UpdateDownloaderService extends Service {
 	private void registerDataListener() {
 		synchronized (mConnectivityManager) {
 			//mTelephonyManager.listen(mDataStateListener, PhoneStateListener.LISTEN_DATA_CONNECTION_STATE);
-			
+
 			registerReceiver(mConnectivityChangesReceiver, mConectivityManagerIntentFilter);
 			mWaitingForDataConnection = true;
 		}
@@ -318,22 +318,22 @@ public class UpdateDownloaderService extends Service {
 	private void notificateDownloadError() {
 		Resources res = getResources();
 		Intent i = new Intent(this, UpdateProcessInfo.class)
-						.putExtra(UpdateProcessInfo.KEY_REQUEST, UpdateProcessInfo.REQUEST_DOWNLOAD_FAILED);
+		.putExtra(UpdateProcessInfo.KEY_REQUEST, UpdateProcessInfo.REQUEST_DOWNLOAD_FAILED);
 
 		PendingIntent contentIntent = PendingIntent.getActivity(this, 0, i,
-															PendingIntent.FLAG_ONE_SHOT);
-		
+				PendingIntent.FLAG_ONE_SHOT);
+
 		Notification notification = new Notification(android.R.drawable.stat_notify_error,
-												res.getString(R.string.not_update_download_error_ticker),
-												System.currentTimeMillis());
-		
+				res.getString(R.string.not_update_download_error_ticker),
+				System.currentTimeMillis());
+
 		notification.flags = Notification.FLAG_AUTO_CANCEL;
 		notification.setLatestEventInfo(
-							this,
-							res.getString(R.string.not_update_download_error_title),
-							res.getString(R.string.not_update_download_error_body),
-							contentIntent);
-		
+				this,
+				res.getString(R.string.not_update_download_error_title),
+				res.getString(R.string.not_update_download_error_body),
+				contentIntent);
+
 		Uri notificationRingtone = Preferences.getPreferences(this).getConfiguredRingtone();
 		if(notificationRingtone == null) {
 			notification.defaults = Notification.DEFAULT_ALL;
@@ -341,19 +341,19 @@ public class UpdateDownloaderService extends Service {
 			notification.defaults = Notification.DEFAULT_LIGHTS | Notification.DEFAULT_VIBRATE;
 			notification.sound = notificationRingtone;
 		}
-		
+
 		//Use a resourceId as an unique identifier
 		mNM.notify(R.string.not_update_download_error_title, notification);
 	}
-	
+
 	private File downloadFile(UpdateInfo updateInfo) {
-		
+
 		HttpClient httpClient = mHttpClient;
 		HttpClient MD5httpClient = mMD5HttpClient;
-		
+
 		HttpUriRequest req, md5req;
 		HttpResponse response, md5response;
-		
+
 		List<URI> updateMirrors = updateInfo.updateFileUris;
 		int size = updateMirrors.size();
 		int start = mRandom.nextInt(size);
@@ -362,75 +362,82 @@ public class UpdateDownloaderService extends Service {
 		for(int i = 0; i < size; i++) {
 			updateURI = updateMirrors.get((start + i)% size);
 			mMirrorName = updateURI.getHost();
-			
+
 			mFileName = updateInfo.fileName; 
 			if (null == mFileName || mFileName.length() < 1) {
 				mFileName = "update.zip";
 			}
 			Log.d(TAG, "mFileName: " + mFileName);
 			
+			boolean md5Available = true;
+
 			mMirrorNameUpdated = false;
 			//mUpdateProcessInfo.updateDownloadMirror(updateURI.getHost());
 			try {
 				req = new HttpGet(updateURI);
 				md5req = new HttpGet(updateURI+".md5sum");
-				
+
 				// Add no-cache Header, so the File gets downloaded each time
 				req.addHeader("Cache-Control", "no-cache");
 				md5req.addHeader("Cache-Control", "no-cache");
-				
+
 				Log.i(TAG, "Trying to download md5sum file from " + md5req.getURI());
 				md5response = MD5httpClient.execute(md5req);
 				Log.i(TAG, "Trying to download update zip from " + req.getURI());
 				response = httpClient.execute(req);
-				
+
 				int serverResponse = response.getStatusLine().getStatusCode();
 				int md5serverResponse = md5response.getStatusLine().getStatusCode();
 				
-				if (serverResponse == 404 || md5serverResponse == 404) {
+
+				if (serverResponse == 404) {
 					Log.e(TAG, "File not found on Server. Trying next one.");
-				} else if(serverResponse != 200 || md5serverResponse != 200) {
+				} else if(serverResponse != 200) {
 					Log.e(TAG, "Server returned status code " + serverResponse + " for update zip trying next mirror");
-					Log.e(TAG, "Server returned status code " + md5serverResponse + " for update zip md5sum trying next mirror");
+
 				} else {
+					if (md5serverResponse != 200) {
+						md5Available = false;
+						Log.e(TAG, "Server returned status code " + md5serverResponse + " for update zip md5sum trying next mirror");
+					}
 					//If directory not exists, create it
 					File directory = new File(Environment.getExternalStorageDirectory()+"/"+mUpdateFolder);
 					if (!directory.exists()) {
 						directory.mkdirs();
 						Log.d(TAG, "UpdateFolder created");
 					}
-					
+
 					mDestinationFile = new File(Environment.getExternalStorageDirectory()+"/"+mUpdateFolder, mFileName);
-					mDestinationMD5File = new File(Environment.getExternalStorageDirectory()+"/"+mUpdateFolder, mFileName + ".md5sum");
-					
 					if(mDestinationFile.exists()) mDestinationFile.delete();
-					if(mDestinationMD5File.exists()) mDestinationMD5File.delete();
-					
-					//getContent can only be Called once. So I have to write the MD5 by hand.
-					//dumpFile(md5response.getEntity(), MD5File);
-					try
-					{
-						Log.i(TAG, "Trying to Read MD5 hash from response");
-						HttpEntity temp = md5response.getEntity();
-						InputStreamReader isr = new InputStreamReader(temp.getContent());
-						BufferedReader br = new BufferedReader(isr);
-						mDownlaodedMD5 = br.readLine().split("  ")[0];
-						Log.d(TAG,"MD5: " + mDownlaodedMD5);
-						br.close();
-						isr.close();
-						
-						if (temp != null)
-							temp.consumeContent();
-						
-						//Write the String in a .md5 File
-						if (mDownlaodedMD5 != null || !mDownlaodedMD5.equals(""))	{
-							writeMD5(mDestinationMD5File, mDownlaodedMD5);
+
+					if (md5Available) {
+						mDestinationMD5File = new File(Environment.getExternalStorageDirectory()+"/"+mUpdateFolder, mFileName + ".md5sum");
+						if(mDestinationMD5File.exists()) mDestinationMD5File.delete();
+
+						try
+						{
+							Log.i(TAG, "Trying to Read MD5 hash from response");
+							HttpEntity temp = md5response.getEntity();
+							InputStreamReader isr = new InputStreamReader(temp.getContent());
+							BufferedReader br = new BufferedReader(isr);
+							mDownlaodedMD5 = br.readLine().split("  ")[0];
+							Log.d(TAG,"MD5: " + mDownlaodedMD5);
+							br.close();
+							isr.close();
+
+							if (temp != null)
+								temp.consumeContent();
+
+							//Write the String in a .md5 File
+							if (mDownlaodedMD5 != null || !mDownlaodedMD5.equals(""))	{
+								writeMD5(mDestinationMD5File, mDownlaodedMD5);
+							}
 						}
-					}
-					catch (Exception e)
-					{
-						Log.e(TAG, "Exception while reading MD5 response: "+e.getMessage());
-						throw new IOException("MD5 Response cannot be read");
+						catch (Exception e)
+						{
+							Log.e(TAG, "Exception while reading MD5 response: "+e.getMessage());
+							throw new IOException("MD5 Response cannot be read");
+						}
 					}
 
 					// Download Update ZIP if md5sum went ok
@@ -438,26 +445,31 @@ public class UpdateDownloaderService extends Service {
 					dumpFile(entity, mDestinationFile);
 					if (entity != null)
 						entity.consumeContent();
+					Log.i(TAG, "Update download finished");
 					
-					Log.i(TAG, "Update download finished. Performing MD5 verification");
-					if(!IOUtils.checkMD5(mDownlaodedMD5, mDestinationFile)) {
-						throw new IOException("MD5 verification failed");
+					if (md5Available) {
+						Log.i(TAG, "Performing MD5 verification");
+						if(!IOUtils.checkMD5(mDownlaodedMD5, mDestinationFile)) {
+							throw new IOException("MD5 verification failed");
+						}
 					}
-					
+
 					//If we reach here, download & MD5 check went fine :)
-				return mDestinationFile;
+					return mDestinationFile;
 				}
 			} catch (Exception ex) {
 				Log.w(TAG, "An error occured while downloading the update file. Trying next mirror", ex);
 			}
 			if(Thread.currentThread().isInterrupted()) break;
 		}
-		
+
 		Log.e(TAG, "Unable to download the update file from any mirror");
-		
+
 		if(mDestinationFile.exists()) mDestinationFile.delete();
-		if(mDestinationMD5File.exists()) mDestinationMD5File.delete();
-		
+		if (null != mDestinationMD5File) {
+			if(mDestinationMD5File.exists()) mDestinationMD5File.delete();
+		}
+
 		return null;
 	}
 
@@ -469,9 +481,9 @@ public class UpdateDownloaderService extends Service {
 			contentLength = 1024;
 		}
 		else Log.i(TAG, "Update size: " + (contentLength/1024) + "KB" );
-		
+
 		long StartTime = System.currentTimeMillis(); 
-		
+
 		//byte[] buff = new byte[64 * 1024];
 		byte[] buff = new byte[(int) contentLength];
 		int read = 0;
@@ -485,7 +497,7 @@ public class UpdateDownloaderService extends Service {
 				totalDownloaded += read;
 				onProgressUpdate(totalDownloaded, (int)contentLength, StartTime);
 			}
-			
+
 			if(read > 0) {
 				throw new IOException("Download Canceled");
 			}
@@ -506,7 +518,7 @@ public class UpdateDownloaderService extends Service {
 	}
 
 	private void writeMD5(File md5File, String md5) throws IOException {
-		
+
 		Log.d(TAG, "Writing the calculated MD5 to disk");
 		FileWriter fw = new FileWriter(md5File);
 		try {
@@ -519,7 +531,7 @@ public class UpdateDownloaderService extends Service {
 
 	private void onProgressUpdate(int downloaded, int total, long StartTime) {
 		if(UPDATE_PROCESS_INFO == null) return;
-		
+
 		if(!mMirrorNameUpdated) {
 			UPDATE_PROCESS_INFO.updateDownloadMirror(mMirrorName);
 			mMirrorNameUpdated = true;
@@ -535,21 +547,21 @@ public class UpdateDownloaderService extends Service {
 			upi.switchToUpdateChooserLayout(null);
 			return;
 		}
-		
+
 		if(UPDATE_PROCESS_INFO == null) {
 			//app is closed, launching a notification
 			Resources res = getResources();
 			Intent i = new Intent(this, ApplyUploadActivity.class)
-								.putExtra(ApplyUploadActivity.KEY_UPDATE_INFO, ui);
-	
+			.putExtra(ApplyUploadActivity.KEY_UPDATE_INFO, ui);
+
 			PendingIntent contentIntent = PendingIntent.getActivity(this, 0, i,
-												PendingIntent.FLAG_CANCEL_CURRENT);
-			
+					PendingIntent.FLAG_CANCEL_CURRENT);
+
 			Notification notification = new Notification(R.drawable.icon_notification,
-												res.getString(R.string.not_update_downloaded_ticker),
-												System.currentTimeMillis());
+					res.getString(R.string.not_update_downloaded_ticker),
+					System.currentTimeMillis());
 			String notificationBody = MessageFormat.format(res.getString(R.string.not_update_downloaded_body),
-												ui.name);
+					ui.name);
 			notification.setLatestEventInfo(this, res.getString(R.string.not_update_downloaded_title), notificationBody, contentIntent);
 			Uri notificationRingtone = Preferences.getPreferences(this).getConfiguredRingtone();
 			if(notificationRingtone == null) {
@@ -558,7 +570,7 @@ public class UpdateDownloaderService extends Service {
 				notification.defaults = Notification.DEFAULT_LIGHTS | Notification.DEFAULT_VIBRATE;
 				notification.sound = notificationRingtone;
 			}
-			
+
 			NotificationManager nm = (NotificationManager) getSystemService(Context.NOTIFICATION_SERVICE);
 			nm.notify(R.string.not_update_downloaded_title, notification);
 		} else {
@@ -569,9 +581,9 @@ public class UpdateDownloaderService extends Service {
 			i.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
 			startActivity(i);
 		}
-		
+
 	}
-	
+
 	public void cancelDownload() {
 		//Thread.currentThread().interrupt();
 		mHandlerThread.interrupt();
